Refactor of machine.rs
- Split machine module for separation of concerns
- Cleaned up help overlay
- Fixes and improvements to README
- Added new categories for crates.io
