+++ Cargo.toml
@@ -9,7 +9,7 @@ homepage = "https://github.com/cenonym/trmt"
 license = "GPL-3.0-or-later"
 readme = "README.md"
 keywords = ["turmite", "turing-machine", "cellular-automaton", "simulation"]
-categories = ["command-line-utilities", "simulation"]
+categories = ["command-line-utilities", "simulation", "science", "visualization", "mathematics" ]
 exclude = ["*.gif", "/.github"]
 
 [dependencies]
