+++ README.md
@@ -8,19 +8,22 @@
     <br>
     <br>
     <img src="/trmt_demo.gif" alt="trmt demo">
+    <p>
+        Demo background by <a href="https://silverwing-vfx.de">Raphael Rau</a>
+    </p>
 </div>
 
 ---
 
 ### About
-[Turmites](https://en.wikipedia.org/wiki/Turmite) are cellular automata where agents, or heads, move on a 2D grid, changing cell states based on rules. Each head reads the current cell, writes a new state, moves in a set direction, then advances.
+**trmt** can simulate multiple turmites simultaneously with customizable rules, colors, and characters, and has support for both [relative and absolute](https://en.wikipedia.org/wiki/Turmite#Relative_vs._absolute_turmites) turmites.
 
-**trmt** simulates multiple turmites simultaneously with customizable rules, colors, and characters, and has support for both [relative and absolute](https://en.wikipedia.org/wiki/Turmite#Relative_vs._absolute_turmites) turmites.
+Running **trmt** will start a simulation with the [default config](#configuration). 
 <br>
 
 ### Features
 - Up to 256 simultaneous heads
-- Support for 16/256-color palettes and true color (RGB/hex)
+- Full color support: 16-color, 256-color, and RGB/hex
 - Deterministic seed-based simulation for reproducible patterns
 - Highly configurable display, controls, and simulation parameters
 - Several rule formats for various degrees of complexity
@@ -30,8 +33,6 @@
 
 ### Installation
 **trmt** only requires [Rust with Cargo](https://www.rust-lang.org/learn/get-started) as a prerequisite.
-> [!NOTE]  
-> Pre-compiled binaries and OS/distro-specific managers will be added.
 
 
 **From crates.io**
@@ -49,11 +50,10 @@ cargo install --path .
 <br>
 
 ### Usage
-Simply run `trmt` in your terminal of choice.
+Simply run `trmt` in your terminal to start a simulation.
 ```bash
 trmt
 ```
-While **trmt** will work in any terminal *(I think)*, the best performance was observed in [Alacritty](https://github.com/alacritty/alacritty) during testing.
 <br>
 
 #### Controls
@@ -93,7 +93,7 @@ seed = ""                           # Empty = random
 
 [display]
 colors = [                          # Array of colors mapped to number of heads sequentially, using hex, RGB or 256-colors.
-    "rgb(241, 113, 54)",          # If there are more heads than colors, remaining colors are generated.
+    "rgb(241, 113, 54)",            # If there are more heads than colors, remaining colors are generated.
     "#45a8e9",
     "229",
 ]
@@ -113,7 +113,7 @@ statusbar = "b"                     # Toggle statusbar overlay
 seed_toggle = "s"                   # Toggle seed
 ```
 > [!TIP]  
-> The seed controls starting position and initial direction of the heads. 
+> The seed stores initial position and direction of the heads.
 
 <br>
 
@@ -121,9 +121,9 @@ seed_toggle = "s"                   # Toggle seed
 > [!NOTE]  
 > A wiki containing a more in-depth guide to the syntax and rule formats will be added to a wiki.
 
-Rules in **trmt** are what defines how the simulation will play out, and how the heads will behave. **trmt** features a custom rule syntax that allows you to create everything from very basic sequential rules, all the way up to academic-level notation *(don't quote me on this)*.
+Rules in **trmt** are what defines how the simulation will play out, and how the heads will behave. **trmt** provides you with tools to simulate everything from very basic sequential rules, all the way up to academic-level notation *(don't quote me on this)*.
 
-A rule consists of several states, and a state holds specific instructions on how a head should move. You can theoretically have several hundred states in a rule, but a lot of the most interesting patterns will appear with just 3-5 states in a rule.
+A rule consists of several states, and a state holds specific instructions on how a head should move when encountered. You can theoretically have several hundred states in a rule, but many of the most interesting patterns will appear with just 3-5 states.
 
 **Available states:**
 - `L/R` - Turn left/right
@@ -135,7 +135,7 @@ A rule consists of several states, and a state holds specific instructions on ho
 The parser has deterministic left-to-right precedence, so `"NE"` always becomes diagonal, never `N+E`.
 
 **Basic sequential format**
-Using the directions above, we can create simple sequential rules that move through each state in turn. One of the simplest and most famous turmites is [Langton's Ant](https://en.wikipedia.org/wiki/Langton%27s_ant), which can be easily replicated with **trmt** using just `RL` as a rule.
+Using the states above, we can create simple sequential rules that move through each state in turn. One of the simplest and most famous turmites is [Langton's Ant](https://en.wikipedia.org/wiki/Langton%27s_ant), which can be replicated with **trmt** using just `RL` as a rule.
 ```toml
 default_rule = "RL"
 ```
@@ -143,17 +143,23 @@ default_rule = "RL"
 Or try something like `WRSWNL` or `RRLL`, or even `RULE` if you're feeling meta.
 
 **Rule operators**
-In addition to these basic directional states, **trmt** also has logical operators for more complex rules that enable internal multi-states and explicit state transitions.
+In addition to these basic directional states, **trmt** also has logical operators for more complex rules:
 
-- `>` - State transition (e.g., R>1 = turn right, go to state 1)
+- `>` - State transition (e.g. R>1 = turn right, go to state 1)
 - `,` - Separates rule combinations for explicit state rules
 - `:` - Separates states in multi-state rules
 
-For example, these operators allow us to explicitly define how our heads transition between states. This makes for very advanced possibilities, and lets us directly translate a traditional turmite rule like `{{{1, 8, 1}, {1, 8, 1}}, {{1, 2, 1}, {0, 1, 0}}}` (a generalized Langton's Ant), into a more ~~opinionated~~ readable syntax:
+For precise control, you can specify which cell state to write:
+- `L1>1` = turn left, write cell state 1, go to state 1
+- `R0>0` = turn right, write cell state 0, go to state 0
+
+This lets us translate a traditional turmite notation like `{{{1, 8, 1}, {1, 8, 1}}, {{1, 2, 1}, {0, 1, 0}}}` into a more ~~opinionated~~ readable syntax:
 ```toml
 default_rule = "L1>1,L1>1:R1>1,D0>0"
 ```
-This results in a rule that constructs a [Fibonacci spiral](https://en.wikipedia.org/wiki/Turmite#/media/File:Turmite-181181121010-10211.svg).
+Which constructs a [Fibonacci spiral](https://en.wikipedia.org/wiki/Turmite#/media/File:Turmite-181181121010-10211.svg).
+
+When building new rules, it is recommended to use `1` head for testing to make the simulation less chaotic.
 <br>
 
 ### Planned
@@ -165,7 +171,7 @@ This results in a rule that constructs a [Fibonacci spiral](https://en.wikipedia
 - [ ] Redesign help and statusbar TUI
 - [ ] Support for 32-bit colors
 - [ ] Proper wiki/documentation
-
+<br>
 
 ### Contributing
 You are very welcome to contribute to the project, be that through feature requests, improvements to the code or by adding functions. Please follow these steps:
